Allow using `git::URL` without `git` feature active
